refactor(architecture): implement class-based modular architecture
- Extract core functionality into dedicated manager classes:
  - APIKeyManager for API key operations
  - CommitMessageGenerator for message generation logic
  - ConfigManager for configuration handling
  - GitManager for Git operations
- Move API key management logic from extension.ts to APIKeyManager
- Create CommitMessageGenerator class to handle prompt generation and API calls
- Implement ConfigManager for centralised configuration access
- Add GitManager to encapsulate Git-related operations
- Update extension.ts to use new manager classes
- Improve error handling and logging across components
